Do not allow requests with the reserved namespace
This commit prevents user requests using the reserved namespace
in order to eliminate the possibility of naming collisions.

The implementation is a little funky because we have to bypass
the check the first time we get the customer namespace (because
we have not yet created the HttpReq so can't send an error back).

BugzID: 43618